# Optimize incremental builds on CI