--- a/.github/scripts/archive_derived_data.sh
+++ b/.github/scripts/archive_derived_data.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+set -e
+
+mkdir -p .github/cache/derived-data-tar-cache
+tar cfPp .github/cache/derived-data-tar-cache/derived-data.tar --format=posix .github/cache/derived-data