--- a/.github/scripts/run_tests.sh
+++ b/.github/scripts/run_tests.sh
@@ -3,6 +3,8 @@
 set -e -u -o pipefail
 
 xcodebuild \
+    -IgnoreFileSystemDeviceInodeChanges=1 \
+    -derivedDataPath .github/cache/derived-data \
     -workspace 'Thumbprint.xcworkspace' \
     -scheme "$scheme" \
     -destination "$destination" \