Cache Derived Data and set the proper values for file modification
times so that xcodebuild can optimize incremental builds.