@kevinmbeaulieu 2021-03-24
From a few runs, here are the durations of the Xcode-related steps ("Build and test" for baseline, "Build and test" + the new optimization steps for variant):

Baseline:
- Run 1: 4min 3s
- Run 2: 3min 28s
- Run 3: 3min 7 s

With optimization:
- Run 1: 3min 54s
- Run 2: 5min 37s
- Run 3: 6min 38s

So looks like it's actually slower with these changes (maybe the way the build system uses derived data is optimized for larger projects 🤷‍♂️). Closing this PR out